PR: Melhorias + Nova arquitetura

+++ src/main/java/com/dtplan/domain/refeicao/Refeicao.java
@@ -2,10 +2,8 @@ package com.dtplan.domain.refeicao;
 
 import com.dtplan.domain.alimento.Alimento;
 import com.dtplan.domain.dieta.Dieta;
-import com.dtplan.domain.dieta.dto.EditarDietaDTO;
-import com.dtplan.domain.refeicao.dto.EditarRefeicaoDTO;
+import com.dtplan.domain.refeicaoAlimento.RefeicaoAlimento;
 import jakarta.persistence.*;
-import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -29,11 +27,11 @@ public class Refeicao {
     private Long id;
     private String descricao;
 
-    private Float calorias; // kcal // Energia:
-    private Float proteina; // g
-    private Float gordura; // lipídeos // g
-    private Float carboidrato; // g
-    private Float fibraAlimentar; // g
+    private double calorias; // kcal // Energia:
+    private double proteina; // g
+    private double gordura; // lipídeos // g
+    private double carboidrato; // g
+    private double fibraAlimentar; // g
 
     @ManyToOne
     @JoinColumn(name = "dieta_id")
@@ -57,7 +55,7 @@ public class Refeicao {
     }
 
     // Método para adicionar um alimento com quantidade à refeição
-    public void adicionarAlimento(Refeicao refeicao, Alimento alimento, Float quantidade) {
+    public void adicionarAlimento(Refeicao refeicao, Alimento alimento, double quantidade) {
         Optional<RefeicaoAlimento> existente = refeicao.getRefeicaoAlimentos().stream()
                 .filter(ra -> ra.getAlimento().getId().equals(alimento.getId()))
                 .findFirst();
@@ -70,7 +68,7 @@ public class Refeicao {
         atualizarNutrientesRefeicao();
     }
 
-    public void editarAlimento(Refeicao refeicao, Long id, Float novaQuantidade) {
+    public void editarAlimento(Refeicao refeicao, Long id, double novaQuantidade) {
         // Localiza o RefeicaoAlimento pelo ID do alimento e atualiza a quantidade
         refeicao.getRefeicaoAlimentos().stream()
                 .filter(ra -> ra.getAlimento().getId().equals(id))
@@ -93,7 +91,7 @@ public class Refeicao {
 
         for (RefeicaoAlimento refeicaoAlimento : refeicaoAlimentos) {
             Alimento alimento = refeicaoAlimento.getAlimento();
-            Float quantidade = refeicaoAlimento.getQuantidade();
+            double quantidade = refeicaoAlimento.getQuantidade();
             calorias += alimento.getCalorias() * quantidade;
             proteina += alimento.getProteina() * quantidade;
             gordura += alimento.getGordura() * quantidade;
